subscribed sticky on theme change. observe watching placed to service.
